requirements.txt
pytorch
tensorflow
transformers
nltk
allennlp
allennlp-models
spacy
spacy-transformers
numpy
datasets